# api/events/api.openapi.yaml
openapi: 3.0.3
info:
  description: Hotels Server API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Hotels Server API
  version: "0.1"
  contact:
    name: Hotels Server API

servers:
  - url: http://localhost:8080/
    description: Hotels Server API

components:
  schemas:

    Event:
      description: Event
      type: object
      properties:
        id:
          description: ID
          type: integer
        owner:
          description: Кто создал
          type: string
        created_at:
          description: Когда создал
          type: string
          format: date
        name:
          description: Название
          type: string
        description:
          description: Описание
          type: string
        when_occurs:
          description: Когда произойдет
          type: string
          format: date
        when_remind:
          description: Когда напомнить
          type: string
          format: date
    Error:
      description: Сообщение об ошибках
      type: object
      properties:
        message:
          type: string

paths:
  /events:
    get:
      summary: Получить массив с информацией о событиях
      operationId: GetEventsList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Создать новое событие
      operationId: CreateNewEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner:
                  description: Кто создал
                  type: string
                created_at:
                  description: Когда создал
                  type: string
                  format: date
                name:
                  description: Название
                  type: string
                description:
                  description: Описание
                  type: string
                when_occurs:
                  description: Когда произойдет
                  type: string
                  format: date
                when_remind:
                  description: Когда напомнить
                  type: string
                  format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Изменение информации о событии
      operationId: EditEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  description: ID
                  type: integer
                owner:
                  description: Кто создал
                  type: string
                created_at:
                  description: Когда создал
                  type: string
                  format: date
                name:
                  description: Название
                  type: string
                description:
                  description: Описание
                  type: string
                when_occurs:
                  description: Когда произойдет
                  type: string
                  format: date
                when_remind:
                  description: Когда напомнить
                  type: string
                  format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Удаление события
      operationId: DeleteEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  description: ID
                  type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"